requirements.txt
streamlit
SQLAlchemy==2.0.27
pymysql
mysqlclient
mysql.connector
